Move Quote/Splice retyping into ReTyper
